refactor: Navigation 구현체가 아닌 인터페이스에 의존하도록 변경
